requirements.txt
gensim>=4.4.0
numpy>=1.24
scipy>=1.10
hnswlib>=0.7
faiss-cpu>=1.7
matplotlib>=3.7
